move session service to slack services folder
